improve (j-jump-prev-register), (j-jump-next-register)
if the current point marker is equal to the marker of the previous register,
the register before the previous register is used as the previous register,
and then try agiain until the above condition is not true.
after finding the appropriate register as the previous register,
jump to it.
(j-jump-next-register) is similar to (j-jump-prev-register).